# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Event-driven multi-channel notification system",
        "title": "Notification System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/metrics": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Get real-time system metrics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/notifications": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "List notifications with filtering and pagination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by channel",
                        "name": "channel",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by batch ID",
                        "name": "batch_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created after (RFC3339)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created before (RFC3339)",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (max 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaginatedResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Create a notification",
                "parameters": [
                    {
                        "description": "Notification",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Notification"
                        }
                    }
                }
            }
        },
        "/api/v1/notifications/batch": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Create batch notifications (up to 1000)",
                "parameters": [
                    {
                        "description": "Batch",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Batch"
                        }
                    }
                }
            }
        },
        "/api/v1/notifications/batch/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get batch status by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Batch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Batch"
                        }
                    }
                }
            }
        },
        "/api/v1/notifications/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get notification by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Notification"
                        }
                    }
                }
            }
        },
        "/api/v1/notifications/{id}/cancel": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Cancel a pending/queued notification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/templates": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "List all templates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Template"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Create a message template",
                "parameters": [
                    {
                        "description": "Template",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Template"
                        }
                    }
                }
            }
        },
        "/api/v1/templates/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Get template by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Template"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check endpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/ws/notifications": {
            "get": {
                "tags": [
                    "system"
                ],
                "summary": "WebSocket endpoint for real-time notification status",
                "responses": {}
            }
        }
    },
    "definitions": {
        "handlers.CreateTemplateRequest": {
            "type": "object",
            "properties": {
                "channel": {
                    "$ref": "#/definitions/models.Channel"
                },
                "content": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Batch": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "delivered": {
                    "type": "integer"
                },
                "failed": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "pending": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "models.Channel": {
            "type": "string",
            "enum": [
                "sms",
                "email",
                "push"
            ],
            "x-enum-varnames": [
                "ChannelSMS",
                "ChannelEmail",
                "ChannelPush"
            ]
        },
        "models.CreateBatchRequest": {
            "type": "object",
            "required": [
                "notifications"
            ],
            "properties": {
                "notifications": {
                    "type": "array",
                    "maxItems": 1000,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/models.CreateNotificationRequest"
                    }
                }
            }
        },
        "models.CreateNotificationRequest": {
            "type": "object",
            "required": [
                "channel",
                "content",
                "recipient"
            ],
            "properties": {
                "channel": {
                    "enum": [
                        "sms",
                        "email",
                        "push"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Channel"
                        }
                    ]
                },
                "content": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "priority": {
                    "enum": [
                        "high",
                        "normal",
                        "low"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Priority"
                        }
                    ]
                },
                "recipient": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "template_id": {
                    "type": "string"
                },
                "template_vars": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "models.Notification": {
            "type": "object",
            "properties": {
                "batch_id": {
                    "type": "string"
                },
                "channel": {
                    "$ref": "#/definitions/models.Channel"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "max_retries": {
                    "type": "integer"
                },
                "priority": {
                    "$ref": "#/definitions/models.Priority"
                },
                "provider_message_id": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "retry_count": {
                    "type": "integer"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "sent_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.Status"
                },
                "template_id": {
                    "type": "string"
                },
                "template_vars": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "pages": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "models.Priority": {
            "type": "string",
            "enum": [
                "high",
                "normal",
                "low"
            ],
            "x-enum-varnames": [
                "PriorityHigh",
                "PriorityNormal",
                "PriorityLow"
            ]
        },
        "models.Status": {
            "type": "string",
            "enum": [
                "pending",
                "queued",
                "processing",
                "delivered",
                "failed",
                "cancelled",
                "scheduled"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusQueued",
                "StatusProcessing",
                "StatusDelivered",
                "StatusFailed",
                "StatusCancelled",
                "StatusScheduled"
            ]
        },
        "models.Template": {
            "type": "object",
            "properties": {
                "channel": {
                    "$ref": "#/definitions/models.Channel"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}